backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.2
python-multipart==0.0.9
openai==1.30.1
PyPDF2==3.0.1
python-docx==1.1.2
docx2txt==0.8
pytesseract==0.3.10
Pillow==10.3.0
